Swapped KordiTextField to regular TextFormField with proper styling in entire app
