+++ lib/sign_in/pages/sign_in_page.dart
@@ -4,7 +4,6 @@ import 'package:go_router/go_router.dart';
 import 'package:kordi_mobile/core/navigation/kordi_router.dart';
 import 'package:kordi_mobile/core/utils/kordi_dialog.dart';
 import 'package:kordi_mobile/core/utils/kordi_flushbar.dart';
-import 'package:kordi_mobile/core/widgets/kordi_text_field.dart';
 import 'package:kordi_mobile/dependency_injection.dart';
 import 'package:kordi_mobile/gen/assets.gen.dart';
 import 'package:kordi_mobile/gen/l10n.dart';
@@ -13,14 +12,13 @@ import 'package:kordi_mobile/sign_in/controllers/sign_in/sign_in_bloc.dart';
 import 'package:kordi_mobile/sign_in/controllers/sign_in_form/sign_in_form_bloc.dart';
 import 'package:kordi_mobile/user/controllers/get_user_cubit.dart';
 
+part 'package:kordi_mobile/sign_in/widgets/sign_in_page_form.dart';
+
 class SignInPage extends StatelessWidget {
   const SignInPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    final ColorScheme colorScheme = theme.colorScheme;
-
     return GestureDetector(
       onTap: () => FocusScope.of(context).unfocus(),
       child: MultiBlocProvider(
@@ -48,135 +46,9 @@ class SignInPage extends StatelessWidget {
               );
             }
           },
-          child: BlocBuilder<SignInFormBloc, SignInFormState>(
-            builder: (context, state) {
-              final signInFormBloc = context.read<SignInFormBloc>();
-              return Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 16),
-                child: Center(
-                  child: SingleChildScrollView(
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Container(
-                          padding: const EdgeInsets.all(16),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(16),
-                            color: colorScheme.primaryContainer,
-                          ),
-                          child: Column(
-                            children: [
-                              Assets.images.welcoming.svg(
-                                width: MediaQuery.of(context).size.width * 0.3,
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(8),
-                                child: KordiTextField(
-                                  labelText: S
-                                      .current.signInPageUsernameLabelTextField,
-                                  hintText:
-                                      S.current.signInPageUsernameHintTextField,
-                                  shouldShowErrorText: state.showUsernameError,
-                                  errorText: S
-                                      .current.signInPageUsernameErrorTextField,
-                                  prefixIcon: Icon(Icons.person),
-                                  onChanged: (username) => signInFormBloc.add(
-                                    SignInFormEvent.updateUsername(username),
-                                  ),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(8),
-                                child: KordiTextField(
-                                  obscureNeeded: true,
-                                  shouldShowErrorText: state.showPasswordError,
-                                  labelText: S
-                                      .current.signInPagePasswordLabelTextField,
-                                  hintText:
-                                      S.current.signInPagePasswordHintTextField,
-                                  errorText: S
-                                      .current.signInPagePasswordErrorTextField,
-                                  prefixIcon: Icon(Icons.lock),
-                                  onChanged: (password) => signInFormBloc.add(
-                                    SignInFormEvent.updatePassword(password),
-                                  ),
-                                ),
-                              ),
-                              ElevatedButton(
-                                style: ButtonStyle(
-                                  minimumSize: MaterialStateProperty.all(
-                                    Size(
-                                      MediaQuery.of(context).size.width * 0.65,
-                                      50,
-                                    ),
-                                  ),
-                                ),
-                                onPressed: () {
-                                  _onSignInButtonClicked(context);
-                                },
-                                child: Builder(
-                                  builder: (context) {
-                                    if (state.isLoading) {
-                                      return Center(
-                                        child: CircularProgressIndicator(),
-                                      );
-                                    }
-                                    return Text(
-                                      S.current.signInPageSignInButtonLabel,
-                                    );
-                                  },
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(top: 8),
-                                child: TextButton(
-                                  onPressed: () {
-                                    SignUpPageRoute().go(context);
-                                  },
-                                  child: Text.rich(
-                                    TextSpan(
-                                      text: S
-                                          .current.signInPageSignUpButtonLabel1,
-                                      children: [
-                                        TextSpan(
-                                          text: S.current
-                                              .signInPageSignUpButtonLabel2,
-                                          style: TextStyle(
-                                            fontWeight: FontWeight.bold,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              );
-            },
-          ),
+          child: _SignInPageForm(),
         ),
       ),
-    );
-  }
-
-  void _onSignInButtonClicked(BuildContext context) {
-    final signInFormBloc = context.read<SignInFormBloc>();
-    final state = signInFormBloc.state;
-    FocusScope.of(context).unfocus();
-    if (!state.isFormValid) {
-      signInFormBloc.add(
-        SignInFormEvent.validateFields(),
-      );
-      return;
-    }
-    signInFormBloc.add(
-      SignInFormEvent.signIn(),
     );
   }
 }
